+++ openfisca_pf/parameters/dicp/impot_foncier/degrevement.yaml
@@ -1,4 +1,4 @@
-premier_degrevement:
+premier:
   description: "Premier dégrèvement appliqué à la valeur locative pour calculer la base imposable de l'impôt foncier"
   metadata:
     type: float
@@ -7,7 +7,7 @@ premier_degrevement:
     2021-12-27:
       value: 0.25
 
-second_degrevement_si_non_loue:
+second_si_non_loue:
   description: "Second dégrèvement appliqué pour calculer la base imposable de l'impôt foncier si le local n'est pas loué"
   metadata:
     type: float
@@ -16,7 +16,7 @@ second_degrevement_si_non_loue:
     2021-12-27:
       value: 0
 
-second_degrevement_si_loue_meuble:
+second_si_loue_meuble:
   description: "Second dégrèvement appliqué pour calculer la base imposable de l'impôt foncier si le local est loué en meublé"
   metadata:
     type: float
@@ -25,7 +25,7 @@ second_degrevement_si_loue_meuble:
     2021-12-27:
       value: 0.30
 
-second_degrevement_si_loue_non_meuble:
+second_si_loue_non_meuble:
   description: "Second dégrèvement appliqué pour calculer la base imposable de l'impôt foncier si le local est loué en non-meublé"
   metadata:
     type: float
@@ -34,7 +34,7 @@ second_degrevement_si_loue_non_meuble:
     2021-12-27:
       value: 0.25
 
-degrevement_pour_baisse_de_revenus_loue_en_meuble_de_tourisme:
+pour_baisse_de_revenus_loue_en_meuble_de_tourisme:
   description: "Dégrèvement sur demande dans le cas d'une baisse de revenus d'un local loué en meublé de tourisme"
   metadata:
     type: float
